+++ plone/app/widgets/testing.py
@@ -9,6 +9,7 @@ from plone.app.testing.layers import IntegrationTesting
 from plone.app.widgets.interfaces import IWidgetsLayer
 from plone.testing import z2
 from zope.configuration import xmlconfig
+from zope.interface import directlyProvides
 from zope.interface import implements
 from zope.publisher.browser import TestRequest as BaseTestRequest
 from zope.schema.interfaces import IVocabularyFactory
@@ -17,17 +18,30 @@ from zope.schema.vocabulary import SimpleVocabulary
 
 
 class ExampleVocabulary(object):
-
     implements(IVocabularyFactory)
 
-    def __call__(self, context):
+    def __call__(self, context, query=None):
+        items = [u'One', u'Two', u'Three']
         tmp = SimpleVocabulary([
-            SimpleTerm('one', 'one', u'One'),
-            SimpleTerm('two', 'two', u'Two'),
-            SimpleTerm('three', 'three', u'Three'),
+            SimpleTerm(it.lower(), it.lower(), it)
+            for it in items
+            if query is None
+            or query.lower() in it.lower()
         ])
         tmp.test = 1
         return tmp
+
+
+def ExampleFunctionVocabulary(context, query=None):
+    items = [u'First', u'Second', u'Third']
+    tmp = SimpleVocabulary([
+        SimpleTerm(it.lower(), it.lower(), it)
+        for it in items
+        if query is None
+        or query.lower() in it.lower()
+    ])
+    return tmp
+directlyProvides(ExampleFunctionVocabulary, IVocabularyFactory)
 
 
 class TestRequest(BaseTestRequest):